requirements.txt
anyio==3.6.2
asgiref==3.4.1
asn1crypto==0.24.0
attrs==25.3.0
Charm-Crypto==0.50
click==8.0.4
contextlib2==21.6.0
contextvars==2.4
cryptography==2.1.4
dataclasses==0.8
exceptiongroup==1.2.2
fastapi==0.83.0
h11==0.13.0
hypothesis==6.130.6
idna==3.10
immutables==0.19
importlib-metadata==4.8.3
keyring==10.6.0
keyrings.alt==3.0
pycrypto==2.6.1
pydantic==1.9.2
pygobject==3.26.1
pyparsing==2.4.0
pyxdg==0.25
SecretStorage==2.3.1
six==1.11.0
sniffio==1.2.0
sortedcontainers==2.4.0
starlette==0.19.1
typing-extensions==4.1.1
uvicorn==0.14.0
zipp==3.6.0
